reporting-analytics/scripts/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
psycopg2-binary>=2.9.3
sqlalchemy>=1.4.0
requests>=2.26.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
statsmodels>=0.13.0
prophet>=1.0.0
python-dateutil>=2.8.0
pytz>=2021.1
